+++ src/lgfx/v1/platforms/esp32/Panel_CVBS.cpp
@@ -93,7 +93,7 @@ namespace lgfx
       _push_idx = 0;
       _using_idx = cache_num - 1;
       prev_index = 0;
-      if ((uint32_t)task_pinned_core >= portNUM_PROCESSORS)
+      if (task_pinned_core >= portNUM_PROCESSORS)
       {
         task_pinned_core = (xPortGetCoreID() + 1) % portNUM_PROCESSORS;
       }
